Rename ConnectDevicePage to ExtensionRequestLedgerPermissionPopup
