+++ extension/src/ExtensionRequestLedgerPermissionPopup/ExtensionRequestLedgerPermissionPopup.tsx
@@ -13,8 +13,8 @@ import { AlertBox } from 'app/components/AlertBox'
 import { WalletErrors } from 'types/errors'
 import { importAccountsActions } from 'app/state/importaccounts'
 import { requestDevice } from 'app/lib/ledger'
-import logotype from '../../../../public/Icon Blue 192.png'
 import { WalletType } from '../../state/wallet/types'
+import logotype from '../../../public/Icon Blue 192.png'
 
 type ConnectionStatus = 'connected' | 'disconnected' | 'connecting' | 'error'
 type ConnectionStatusIconPros = {
@@ -44,7 +44,7 @@ function ConnectionStatusIcon({ success = true, label, withMargin = false }: Con
   )
 }
 
-export function ConnectDevicePage() {
+export function ExtensionRequestLedgerPermissionPopup() {
   const { t } = useTranslation()
   const dispatch = useDispatch()
   const [connection, setConnection] = useState<ConnectionStatus>('disconnected')
